# BoltDB wrapper nano optimization which makes code a bit prettier too